Fixing bugs and other issues caught by Sonar (#377)
* Fixing bugs from Sonar

* Clean up SONAR code smells

+++ modules/dsp/chowdsp_dsp_utils/Convolution/chowdsp_ConvolutionEngine.cpp
@@ -59,7 +59,7 @@ template <typename FFTEngineType>
 void ConvolutionEngine<FFTEngineType>::updateSegmentsIfNecessary (size_t numSegmentsToUpdate, std::vector<juce::AudioBuffer<float>>& segments, size_t fftSize)
 {
     if (numSegmentsToUpdate == 0
-        || numSegmentsToUpdate != (size_t) segments.size()
+        || numSegmentsToUpdate != segments.size()
         || (size_t) segments[0].getNumSamples() != fftSize * 2)
     {
         segments.clear();
